fix: 修复测试和文档，优化代码质量
- 修复 SoAStorage 索引逻辑，使用索引作为 ID
- 修复字符串池空值处理
- 更新测试代码以匹配新实现
- 修复所有文档测试
- 修复 cargo clippy 警告
- 格式化代码
- 版本更新至 0.1.1

+++ src/symbol/link.rs
@@ -447,7 +447,7 @@ mod tests {
                 .link(
                     SymbolId::new(i),
                     SymbolId::new(i + 1),
-                    DefinitionLocation::new(1, i as u32 * 10),
+                    DefinitionLocation::new(1, i * 10),
                 )
                 .unwrap();
         }
